backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pydantic-settings==2.5.2
pydantic[email]==2.9.2
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
python-multipart==0.0.12
slowapi==0.1.9
python-docx==1.1.2
openpyxl==3.1.5
